Parallel a little differently(actually use all dynos)

+++ app/jobs/build_heatmap_job.rb
@@ -1,8 +1,18 @@
+def abs_ceil(num)
+  num >= 0 ? num.ceil(1) : num.floor(1)
+end
+
+def abs_floor(num)
+  num >= 0 ? num.floor(1) : num.ceil(1)
+end
+
 class BuildHeatmapJob < ApplicationJob
   queue_as :build_heatmap
   sidekiq_options retry: 0
 
   NUM_SEGMENTS=(ENV['NUM_HEATMAP_THREADS'] || 8).to_i
+  STEP_PRECISION=3
+  STEP=(0.1**STEP_PRECISION).round(STEP_PRECISION) # 0.001
 
   def perform(build_status, job_retry=false)
     begin
@@ -11,22 +21,28 @@ class BuildHeatmapJob < ApplicationJob
       job_retry ||= build_status.created_at < 15.minutes.ago
       build_status.update!(state:'received', percent:100)
 
-      unless job_retry
-        HeatmapPoint.delete_all
-      end
+      HeatmapPoint.delete_all if build_status.rebuild? && !job_retry
       
       build_status.update!(state:'branching', percent:100)
 
-      (1..NUM_SEGMENTS).each do |segment|
-        if build_status.build_heatmap_segment_statuses.where(segment:segment).none?
-          build_segment_status = build_status.build_heatmap_segment_statuses.create(state:'initialized', percent:100, segment:segment)
-          BuildHeatmapSegmentJob.perform_later(build_segment_status, job_retry)
+      south_west = furthest_south_west_local
+      north_east = furthest_north_east_local
+      lat = south_west[0]
+      if build_status.build_heatmap_segment_statuses.none?
+        (1..NUM_SEGMENTS).each do |segment|
+          build_segment_status = build_status.build_heatmap_segment_statuses.create(state:'initialized', percent:100, segment:segment, current_lat:lat)
+          BuildHeatmapSegmentJob.perform_later(build_segment_status)
+          lat = (lat+STEP).round(STEP_PRECISION)
         end
       end
+      build_status.update!(current_lat:(lat-STEP).round(STEP_PRECISION))
+
+      # in case its different due to an old build and a change in num_segments
+      num_segments_this_build = build_status.build_heatmap_segment_statuses.count
 
       until build_status.reload.build_heatmap_segment_statuses.all?(&:atleast_isochrones_state?)
         sleep(5)
-        build_status.update!(percent:build_status.build_heatmap_segment_statuses.count(&:atleast_isochrones_state?)*100/NUM_SEGMENTS, updated_at:Time.now)
+        build_status.update!(percent:build_status.build_heatmap_segment_statuses.count(&:atleast_isochrones_state?)*100/num_segments_this_build, updated_at:Time.now)
       end
 
       return if error_found(build_status)
@@ -35,7 +51,7 @@ class BuildHeatmapJob < ApplicationJob
 
       until build_status.reload.build_heatmap_segment_statuses.reload.all?(&:atleast_isochrones_complete_state?)
         sleep(5)
-        build_status.update!(percent:(build_status.build_heatmap_segment_statuses.sum { |segment_status| segment_status.percent/NUM_SEGMENTS }).round(3), updated_at:Time.now)
+        build_status.update!(percent:(build_status.build_heatmap_segment_statuses.sum { |segment_status| segment_status.percent/num_segments_this_build }).round(3), updated_at:Time.now)
       end
 
       return if error_found(build_status)
@@ -44,7 +60,7 @@ class BuildHeatmapJob < ApplicationJob
 
       until build_status.reload.build_heatmap_segment_statuses.all? { |segment_status| segment_status.error ||  segment_status.state == 'complete' } 
         sleep(5)
-        build_status.update!(percent:(build_status.build_heatmap_segment_statuses.sum { |segment_status| segment_status.percent/NUM_SEGMENTS }).round(3), updated_at:Time.now)
+        build_status.update!(percent:calc_total_heatmap_percent(build_status, num_segments_this_build, south_west, north_east), updated_at:Time.now)
       end
       return if error_found(build_status)
       build_status.update!(percent:100, state:'complete')
@@ -53,7 +69,23 @@ class BuildHeatmapJob < ApplicationJob
     end
   end
 
+  def self.furthest_south_west
+    [abs_floor(GroceryStore.minimum(:lat)-0.3), abs_floor(GroceryStore.minimum(:long))-0.3]
+  end
+
+  def self.furthest_north_east
+    [abs_ceil(GroceryStore.maximum(:lat)+0.3), abs_ceil(GroceryStore.maximum(:long)+0.3)]
+  end
+
   private
+
+  def furthest_south_west_local
+    [abs_floor(GroceryStore.minimum(:lat)-0.3), abs_floor(GroceryStore.minimum(:long))-0.3]
+  end
+
+  def furthest_north_east_local
+    [abs_ceil(GroceryStore.maximum(:lat)+0.3), abs_ceil(GroceryStore.maximum(:long)+0.3)]
+  end
 
   def error_found(build_status)
     errored = build_status.build_heatmap_segment_statuses.where.not(error:nil).first
@@ -61,5 +93,13 @@ class BuildHeatmapJob < ApplicationJob
       build_status.update!(error: errored.error)
     end
     errored
+  end
+
+  def calc_total_heatmap_percent(build_status, num_segments, south_west, north_east)
+    long_percent = build_status.build_heatmap_segment_statuses.sum { |segment_status|
+      segment_status.atleast_heatmap_state? ? (segment_status.percent/num_segments) : 0 # 0 if not yet in the right state
+    }
+    lat_percent_per_step = (STEP / (north_east[0]-south_west[0]+STEP))
+    (((build_status.current_lat-south_west[0])/(north_east[0]-south_west[0]+STEP)+long_percent*lat_percent_per_step)*100).round(3)
   end
 end
